added runnable thread for the model
